requirements.txt
alpaca-py>=0.21.0
pandas>=2.1.0
pandas-ta>=0.3.14b
numpy>=1.25.0
pyyaml>=6.0
python-dotenv>=1.0.0
rich>=13.0.0
